Add MarkupBuilder#labelOf
Add '#labelOf' to create a 'label' node with appropriate 'for'
attribute. When 'gender' is given as its first argument, it sets
'field-gender' to the 'for' attribute of 'label' node. If the form has
a name 'alpha', the value of 'for' attribute should be
'alpha-field-gender'.

When the fieldNamePrefix is set by #fieldsFor, the value of 'for'
attribute should has a form like 'field-items-1-gender'.